refactor(Symbol): introduce unified Symbol module
closes #844
