--- src/util/Symbol_dispose.ts
@@ -1,15 +1,0 @@
-import {root} from './root';
-
-if (!root.Symbol) {
-  root.Symbol = {};
-}
-
-if (!root.Symbol.dispose) {
-  if (typeof root.Symbol.for === 'function') {
-    root.Symbol.dispose = root.Symbol.for('dispose');
-  } else {
-    root.Symbol.dispose = '@@dispose';
-  }
-}
-
-export const $$dispose = root.Symbol.dispose;
