+++ src/util/subscribeToResult.ts
@@ -1,7 +1,6 @@
 import {Subscriber} from '../Subscriber';
 import {Observable} from '../Observable';
-import {$$iterator} from '../util/Symbol_iterator';
-import {$$observable} from '../util/Symbol_observable';
+import {SymbolShim} from '../util/SymbolShim';
 import {Subscription} from '../Subscription';
 import {InnerSubscriber} from '../InnerSubscriber';
 import {OuterSubscriber} from '../OuterSubscriber';
@@ -47,7 +46,7 @@ export function subscribeToResult<T, R>(outerSubscriber: OuterSubscriber<T, R>,
       setTimeout(() => { throw err; });
     });
     return destination;
-  } else if (typeof result[$$iterator] === 'function') {
+  } else if (typeof result[SymbolShim.iterator] === 'function') {
     for (let item of result) {
       destination.next(item);
       if (destination.isUnsubscribed) {
@@ -57,8 +56,8 @@ export function subscribeToResult<T, R>(outerSubscriber: OuterSubscriber<T, R>,
     if (!destination.isUnsubscribed) {
       destination.complete();
     }
-  } else if (typeof result[$$observable] === 'function') {
-    const obs = result[$$observable]();
+  } else if (typeof result[SymbolShim.observable] === 'function') {
+    const obs = result[SymbolShim.observable]();
     if (typeof obs.subscribe !== 'function') {
       destination.error('invalid observable');
     } else {
